implement #40

+++ src/cloudscribe.Pagination.Models/PagedResult.cs
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 // Author:					Joe Audette
 // Created:					2017-09-27
-// Last Modified:			2017-09-27
+// Last Modified:			2018-06-16
 //
 
 using System.Collections.Generic;
@@ -16,8 +16,8 @@ namespace cloudscribe.Pagination.Models
             Data = new List<T>();
         }
         public List<T> Data { get; set; }
-        public int TotalItems { get; set; } = 0;
-        public int PageNumber { get; set; } = 1;
+        public long TotalItems { get; set; } = 0;
+        public long PageNumber { get; set; } = 1;
         public int PageSize { get; set; } = 1;
     }
 }
