+++ src/cloudscribe.Web.Pagination/PaginationLink.cs
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 // Author:					Joe Audette
 // Created:					2015-07-06
-// Last Modified:			2015-07-15
+// Last Modified:			2018-06-16
 //
 
 namespace cloudscribe.Web.Pagination
@@ -13,7 +13,7 @@ namespace cloudscribe.Web.Pagination
 
         public bool IsCurrent { get; set; } = false;
 
-        public int PageNumber { get; set; } = -1;
+        public long PageNumber { get; set; } = -1;
 
         public string Text { get; set; } = string.Empty;
