+++ src/cloudscribe.Web.Pagination/PaginationSettings.cs
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 // Author:					Joe Audette
 // Created:					2014-11-15
-// Last Modified:			2016-02-14
+// Last Modified:			2016-06-16
 //
 
 namespace cloudscribe.Web.Pagination
@@ -10,13 +10,13 @@ namespace cloudscribe.Web.Pagination
 
     public class PaginationSettings
     {
-        public int TotalItems { get; set; } = 0;
+        public long TotalItems { get; set; } = 0;
 
         public int ItemsPerPage { get; set; } = 10;
 
-        public int CurrentPage { get; set; } = 1;
+        public long CurrentPage { get; set; } = 1;
 
-        public int MaxPagerItems { get; set; } = 10;
+        public long MaxPagerItems { get; set; } = 10;
 
         public bool ShowFirstLast { get; set; } = false;
